Refactor theme watcher to shared utility
The `initializeThemeWatcher` and `updateTheme` functions were duplicated
between `pages-list.js` and `popup.js`. This commit extracts them into a
single shared file (`shared/theme.js`) and imports the exported
`initializeThemeWatcher` where it was originally used.

This removes roughly 15-20 lines of redundant code and improves
maintainability, keeping theme-related logic consolidated.
