+++ shared/theme.js
@@ -1,0 +1,16 @@
+// Theme change detection and handling
+export function initializeThemeWatcher() {
+  const darkModeQuery = window.matchMedia('(prefers-color-scheme: dark)');
+
+  // Initial theme application
+  updateTheme(darkModeQuery.matches);
+
+  // Detect theme change
+  darkModeQuery.addEventListener('change', (e) => {
+    updateTheme(e.matches);
+  });
+}
+
+function updateTheme(isDark) {
+  document.body.setAttribute('data-theme', isDark ? 'dark' : 'light');
+}
